requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0

# ML/AI dependencies
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78

# Audio for alarms
pygame==2.5.2

# Utilities
python-multipart==0.0.6

